requirements.txt
# Install these requirements if you wish to contribute to this project.

# ===========================================================================
# Packages essential to this app. Needed by anyone who wants to use this app.
# ===========================================================================
django==1.4.3
django-cms==2.3.5
django-filer==0.9.4
South==0.7.6
Pillow==1.7.8


# ==============================================================
# Packages needed for running the tests. Needed by contributors.
# ==============================================================
factory-boy==1.2.0
django-libs==0.14
ipdb==0.7
Fabric
mock
django-nose
coverage
django-coverage
